+++ src/views/core/bodies/CreateMember.vue
@@ -79,7 +79,7 @@
             <span class="icon is-small is-left"><font-awesome-icon icon="fa envelope" /></span>
             <input class="input" type="text" required v-model="user.username" />
           </div>
-          <p class="help is-danger" v-if="errors.name">{{ errors.name.join(', ')}}</p>
+          <p class="help is-danger" v-if="errors.username">{{ errors.username.join(', ')}}</p>
         </div>
 
         <div class="field">
@@ -130,10 +130,10 @@ export default {
       this.isSaving = true
       this.errors = {}
 
-      this.axios.post(this.services['core'] + '/bodies/' + this.$route.params.id + '/create-member', this.user).then((response) => {
+      this.axios.post(this.services['core'] + '/bodies/' + this.$route.params.id + '/create-member', this.user).then(() => {
         this.isSaving = false
         this.$root.showSuccess('User is created.')
-        this.$router.push({ name: 'oms.members.view', params: { id: response.data.data.id } })
+        this.$router.push({ name: 'oms.bodies.members', params: { id: this.$route.params.id } })
       }).catch((err) => {
         this.isSaving = false