+++ src/views/events/Edit.vue
@@ -33,6 +33,16 @@
       </div>
 
       <form @submit.prevent="saveEvent()">
+        <div class="field notification is-warning">
+          <label class="label">COVID-19 restrictions</label>
+          <p>
+            Please follow the advice of local authorities for organising public events!
+          </p><p>
+            Take into account the social distancing requirements and make sure you don't exceed the number of people allowed in the same place. Also, cleaning and sanitary measures must be respected.
+          </p><p>
+            If it is needed, you are allowed to expel participants who don't follow the rules after clear instructions.
+          </p>
+        </div>
         <div class="notification is-info" v-if="!$route.params.id">
           <div class="content">
             <p>If you want to upload the image, please add it after creating the event by going to "Edit event" and uploading it there.</p>
@@ -59,7 +69,7 @@
           <div class="control">
             <textarea class="textarea" placeholder="e.g. Hello world" required v-model="event.description"></textarea>
           </div>
-          <label class="label">Description preview</label>
+          <label class="label">Preview <MarkdownTooltip/></label>
           <div class="content">
             <span v-html="$options.filters.markdown(event.description)" />
           </div>
@@ -94,7 +104,7 @@
         </div>
 
         <div class="notification is-info" v-if="!$route.params.id">
-          Please select the event type wisely. <strong>It can't be changed later.</strong>
+          Please select the event type wisely. <strong>It cannot be changed later.</strong>
         </div>
 
         <div class="field" v-if="!$route.params.id">
@@ -127,6 +137,65 @@
             <input class="input" type="number" v-model="event.max_participants" min="0" @input="$root.nullifyIfEmpty(event, 'max_participants')"/>
           </div>
           <p class="help is-danger" v-if="errors.max_participants">{{ errors.max_participants.join(', ') }}</p>
+        </div>
+        <div class="field notification is-info">
+          <label class="label" style="color: white">Regarding water access</label>
+          <p>
+            Please make sure that the participants have access to a reasonable amount of water during the event.
+          </p>
+          <p>
+            Tip: Advise your participants to bring their own refillable water bottles and remind them throughout the event to refill them.
+          </p>
+        </div>
+
+        <div class="field">
+          <label class="label">Number of meals provided per day <span class="has-text-danger">*</span></label>
+          <div class="control">
+            <div class="control">
+              <input class="input" type="number" v-model="event.meals_per_day" min="0" max="3" required />
+            </div>
+          </div>
+          <p class="help is-danger" v-if="errors.meals_per_day">{{ errors.meals_per_day.join(', ') }}</p>
+        </div>
+        <div class="field">
+          <label class="label">Accommodation type <span class="has-text-danger">*</span></label>
+          <div class="notification is-info" v-if="!$route.params.id">
+            <p>Accommodation can be for instance camping, hostel, hosting by the members of the local, or in a gym. Write "none" for online events.</p>
+          </div>
+          <div class="control">
+            <input class="input"  v-model="event.accommodation_type" required />
+          </div>
+          <p class="help is-danger" v-if="errors.accommodation_type">{{ errors.accommodation_type.join(', ') }}</p>
+        </div>
+
+        <div class="subtitle is-fullwidth has-text-centered">Optional Programme</div>
+        <hr />
+
+        <div class="notification is-info">
+          <div class="content">
+            <p>You may offer an optional feature to your event. If so, please specify the optional programme and its cost. Leave the fields empty if there is no extra fee charged. Be concise in the description: "trip to city X", "ice-skating", "extra museum".</p>
+          </div>
+        </div>
+        <div class="field">
+          <label class="label">Optional Fee</label>
+          <div class="control">
+            <div class="field has-addons">
+              <div class="control">
+                <a class="button is-static">€</a>
+              </div>
+              <div class="control">
+                <input class="input" type="number" v-model="event.optional_fee" min="0" />
+              </div>
+            </div>
+          </div>
+          <p class="help is-danger" v-if="errors.optional_fee">{{ errors.optional_fee.join(', ') }}</p>
+        </div>
+        <div class="field">
+          <label class="label">Optional Programme</label>
+          <div class="control">
+            <input class="input" type="text" v-model="event.optional_programme" />
+          </div>
+          <p class="help is-danger" v-if="errors.optional_programme">{{ errors.optional_programme.join(', ') }}</p>
         </div>
 
         <div class="subtitle is-fullwidth has-text-centered">Event dates</div>
@@ -305,7 +374,7 @@
               <li><b>number</b> - a number (like "How much events in AEGEE I've visited")</li>
               <li><b>select</b> - predefined set of values (like "Vegan", "Vegetarian" and "Meat-eater" for meals type)</li>
               <li><b>checkbox</b> - a yes/no question (like "Do I need a visa?").
-                Combine it with "requred" field to only allow this to be checked in order to to apply (like "I give my consent to share my data with third parties")
+                Combine it with "required" field to only allow this to be checked in order to to apply (like "I give my consent to share my data with third parties")
               </li>
             </ul>
           </div>
@@ -434,6 +503,15 @@
           </div>
         </div>
 
+        <div class="field">
+          <label class="label">How to travel to your country</label>
+          <p>You may provide an optional link to provide useful information about travelling to your country.</p>
+          <div class="control">
+            <input class="input" type="url" v-model="event.link_info_travel_country" />
+          </div>
+          <p class="help is-danger" v-if="errors.link_info_travel_country">{{ errors.link_info_travel_country.join(', ') }}</p>
+        </div>
+
         <div class="subtitle is-fullwidth has-text-centered">EQAC approval fields</div>
         <hr/>
 
@@ -443,6 +521,8 @@
             <p>
               You can omit specifying them when creating the event, but
               <strong> you won't be able to submit event to EQAC if these fields are not set.</strong>
+
+              For online events, you are able to use https://online-event.example.com/ as a placeholder for the budget.
             </p>
             <p>Please provide the link to Google spreadsheets for the event program and budget.</p>
             <p><a href="https://docs.google.com/spreadsheets/u/1/?ftv=1&tgif=d" target="_blank">
@@ -461,7 +541,7 @@
         <div class="field">
           <label class="label">Link to event budget</label>
           <div class="control">
-            <input class="input" type="text" v-model="event.budget" />
+            <input class="input" type="url" v-model="event.budget" />
           </div>
           <p class="help is-danger" v-if="errors.is_budget_set">{{ errors.is_budget_set.join(', ') }}</p>
         </div>
@@ -469,7 +549,7 @@
         <div class="field">
           <label class="label">Link to event program</label>
           <div class="control">
-            <input class="input" type="text" v-model="event.programme" />
+            <input class="input" type="url" v-model="event.programme" />
           </div>
           <p class="help is-danger" v-if="errors.is_programme_set">{{ errors.is_programme_set.join(', ') }}</p>
         </div>
@@ -492,13 +572,15 @@ import { MglMap, MglMarker, MglNavigationControl } from 'vue-mapbox'
 import constants from '../../constants'
 import credentials from '../../credentials'
 import TimezoneNotification from '../../components/notifications/TimezoneNotification'
+import MarkdownTooltip from '../../components/tooltips/MarkdownTooltip'
 
 export default {
   components: {
     MglMap,
     MglMarker,
     MglNavigationControl,
-    TimezoneNotification
+    TimezoneNotification,
+    MarkdownTooltip
   },
   name: 'EditEvent',
   data () {
@@ -519,7 +601,12 @@ export default {
         starts: null,
         ends: null,
         body_id: null,
-        body: null
+        body: null,
+        optional_fee: null,
+        meals_per_day: 0,
+        optional_programme: null,
+        link_info_travel_country: null,
+        accommodation_type: ''
       },
       autoComplete: {
         members: { name: '', values: [], loading: false }